Create Alien models and add them to the scene.
I created an Alien model based off a Space Invaders sprite inside Blender and I added them into the Game. I had to increase the FoV (Field of View) of the Camera to fit more of them on the screen and move the ship farther from the Aliens so the ship isn't very close to the Aliens.