Actually add movement to the Game.
The Aliens finally move. I wrote bound code that gets the bounding box of the alien group and theres code that moves the alien with smoothed lerp movement and it moves down and flips direction if it detects the group is at the edge of the camera according to the current moving direction of the group.