The Aliens can shoot now.
In this commit, I added the ability for the aliens to shoot. The way this works is that the Group GameObject starts a loop where it passes a message down the chain from group, to a random row, and then the random row picks an alien, and then the laser spawns. Once the laser is destroyed, the alien sends a message up the chain to the group, and then it spawns another laser.

I also removed materials from the laser blend as I am doing it in engine this time as its easier to switch colors and I don't have to make a completely brand new model for a color change.