Add a Life System and an Explosion particle when shot.
When you get shot, you explode and the game pauses when the explosion is happening and unpauses after and your character gets teleported to the beginning position.

Right now, theres no Main Menu or End Screen so theres not much to do after you run out of lives, the Game just continues as on.