Fix the Main Menu's controls and work on a GameOver screen.
This commit fixes the Main Menu's UI, finally. I got this working by disabling the raycaster on the LoadingScreen's Canvas and I made a script to make the EventSystem select the dynamically created "Panel Settings" GameObject which is a child of the EventSystem which is required for the controller to work and change focus.