Implement Graphics Settings.
This commit implements the functionality of the Graphics Settings. You can change your resolution scale, your resolution, you can make it windowed, turn on vsync, and more. and it all saves its self to a persisted file. I think there might be some bugs with VSync but I got this implemented which is nice.